requirements.txt
llama-index == 0.11.17
llama-index-llms-gemini == 0.3.7
llama-index-embeddings-gemini == 0.2.1
llama-index-vector-stores-pinecone  == 0.2.1
fastapi == 0.115.0
uvicorn == 0.31.1
python-dotenv == 1.0.1
sentence-transformers == 3.2.0
docx2txt == 0.8
pinecone-client == 5.0.1